Implements Dog image fetch api into componment
