+++ src/components/DogImage.js
@@ -1,20 +1,11 @@
-import { useState } from 'react';
-import defaultDogImage from '../assets/dog.jpg';
-
-const DOG_IMAGE_RANDOMIZER_API = 'https://dog.ceo/api/breeds/image/random';
-
-const Dog = (_props) => {
-  const [dogImage, _setDogImage] = useState(defaultDogImage);
-
-  fetch(DOG_IMAGE_RANDOMIZER_API)
-    .then(response => response.json())
-    .then(data => console.log(data));
+const DogImage = (props) => {
+  const { dogImage } = props;
 
   return (
     <div>
-      <img src={dogImage} alt="Logo" />
+      <img src={dogImage} alt="Random Dog" />
     </div>
   );
 }
 
-export default Dog;
+export default DogImage;
